Revert "Fix example with static_cast to avoid ambiguity"
This reverts commit 35d02b0a45c04b90f43b73216dc3704ec27d6560.

+++ examples/0000-arduino_send_telemetry/0000-arduino_send_telemetry.ino
@@ -178,11 +178,10 @@ void loop() {
   Serial.println("Sending telemetry data...");
 #endif
   // Uploads new telemetry to ThingsBoard using MQTT.
-  // See https://thingsboard.io/docs/reference/mqtt-api/#telemetry-upload-api for more details.
-  // Static_cast, helps the Telemetry constructor which type is wanted exactly, because if this is not given
-  // conversion between float and bool can cause compile time failure with ambiguity which constructor is exactly wanted
-  tb.sendTelemetryData(TEMPERATURE_KEY, static_cast<int>(random(10, 31)));
-  tb.sendTelemetryData(HUMIDITY_KEY, static_cast<int>(random(40, 90)));
+  // See https://thingsboard.io/docs/reference/mqtt-api/#telemetry-upload-api
+  // for more details
+  tb.sendTelemetryData(TEMPERATURE_KEY, random(10, 31));
+  tb.sendTelemetryData(HUMIDITY_KEY, random(40, 90));
 
   tb.loop();
 }
